Allow pool size to be > 2 GB
Change pool-based sizes to be unsigned long long instead of int.

The interface is still limited to size_t (4 byte) amounts, so the only
way to effectively allocate a pool buffer that is > 2 GB in size is to
use CMEM_allocPoolPhys(), which doesn't involve a size parameter nor
need a size back from the driver since it doesn't mmap() the buffer.
When using CMEM_allocPoolPhys(), the user "knows" the pool number that
contains the big buffers and also the size of those buffers.