Allow bigger sizes for CMEM_getBlock and CMEM_getPool
CMEM_getBlock() (and CMEM_getBlockAttrs) was broken for the new extended
blocks.  The interface has been changed to allow for returning a size
greater than the value that will fit in a 'size_t' type.

CMEM_getPool() did not allow a size greater than the value that will fit
in a 'size_t' type.  This API has been changed to allow specifying an
extended size, to support pools with an extended buffer size.