Add support for Android to use mmap64
In the Android Lollipop build environment, the bionic mmap() API doesn't
work for addresses (off_t offset, really) >= 2 GB.  Technically, an off_t
offset can't be >= 2 GB since it's a signed 32-bit type, but it works in
non-Android environments, and a cmemk.c driver module change would need
to be made to accommodate using an offset < 2 GB.

This commit adds internal usage of mmap64 in the CMEM library, plus adds
new 64-bit "flavor" APIs:
    CMEM_getPhys64()
    CMEM_map64()
    CMEM_registerAlloc64()
    CMEM_allocPhys64()
    CMEM_allocPoolPhys64()
    CMEM_freePhys64()
to be used by Android user applications that need to do so, in lieu of
the non-64 counterparts.  An application source file build needs to
    -DLINUXUTILS_BUILDOS_ANDROID
in order for the prototypes of these new APis to be exposed in the header
file cmem.h.