ks2: set dma_pfn_offset for keystone2 devices
On kestone2 SOC where aliased memory is being used dma_alloc_coherent()
cannot be called with pointer to device set to NULL. The real
"struct device" has to be used. And the dma_pfn_offset has to be set
correctly. Otherwise the function returns wrong dma address.

This commit adds the cmem_cma_dev_0 pointer, which is set by
device_create(). It sets the correct coherent_dma_mask and dma_pfn_offset
for this device.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>