cmemk: Introduce cmem_dma_offset_configure wrapper
The Keystone2 platforms require the dma_pfn_offset to be configured
on the CMEM device. Introduce a new cmem_dma_offset_configure() wrapper
function so that the Linux kernel version variations can be localized
to a single function. The wrapper is only effective for ARCH_KEYSTONE
with LPAE enable (no change in functionality) and is a no-op for all
other builds.

Signed-off-by: Suman Anna <s-anna@ti.com>